+++ lib_xua/api/xua_conf_default.h
@@ -1637,3 +1637,12 @@ enum USBEndpointNumber_Out
 #error CODEC_MASTER with DSD is currently unsupported
 #endif
 
+/**
+ * @brief Device interface GUID.
+ *
+ * This is provided as part of the device registry property in the MSOS 2.0 descriptor.
+ * Default: "{89C14132-D389-4FF7-944E-2E33379BB59D}" User can override by defining their own in xua_conf.h
+ */
+#ifndef WINUSB_DEVICE_INTERFACE_GUID
+#define WINUSB_DEVICE_INTERFACE_GUID               "{89C14132-D389-4FF7-944E-2E33379BB59D}"
+#endif