+++ lib_xua/host/xmosdfu/xmosdfu.cpp
@@ -1,4 +1,4 @@
-// Copyright 2012-2021 XMOS LIMITED.
+// Copyright 2012-2024 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
 #include <stdio.h>
 #include <stdlib.h>
@@ -45,9 +45,12 @@ device_pid_t pidList[] = {
 };
 
 unsigned int XMOS_DFU_IF = 0;
+static int dfu_timeout = 5000; // 5s
 
-#define DFU_REQUEST_TO_DEV 0x21
-#define DFU_REQUEST_FROM_DEV 0xa1
+#define USB_BMREQ_H2D_CLASS_INT (LIBUSB_ENDPOINT_OUT | LIBUSB_REQUEST_TYPE_CLASS | LIBUSB_RECIPIENT_INTERFACE)
+#define USB_BMREQ_D2H_CLASS_INT (LIBUSB_ENDPOINT_IN | LIBUSB_REQUEST_TYPE_CLASS | LIBUSB_RECIPIENT_INTERFACE)
+
+#define USB_BMREQ_H2D_VENDOR_INT (LIBUSB_ENDPOINT_OUT | LIBUSB_REQUEST_TYPE_VENDOR | LIBUSB_RECIPIENT_INTERFACE)
 
 // Standard DFU requests
 #define DFU_DETACH 0
@@ -63,8 +66,20 @@ unsigned int XMOS_DFU_IF = 0;
 #define XMOS_DFU_REVERTFACTORY        0xf1
 #define XMOS_DFU_RESETINTODFU         0xf2
 #define XMOS_DFU_RESETFROMDFU         0xf3
-#define XMOS_DFU_SAVESTATE            0xf5
-#define XMOS_DFU_RESTORESTATE         0xf6
+
+enum dfu_state {
+	DFU_STATE_appIDLE		= 0,
+	DFU_STATE_appDETACH		= 1,
+	DFU_STATE_dfuIDLE		= 2,
+	DFU_STATE_dfuDNLOAD_SYNC	= 3,
+	DFU_STATE_dfuDNBUSY		= 4,
+	DFU_STATE_dfuDNLOAD_IDLE	= 5,
+	DFU_STATE_dfuMANIFEST_SYNC	= 6,
+	DFU_STATE_dfuMANIFEST		= 7,
+	DFU_STATE_dfuMANIFEST_WAIT_RST	= 8,
+	DFU_STATE_dfuUPLOAD_IDLE	= 9,
+	DFU_STATE_dfuERROR		= 10
+};
 
 static libusb_device_handle *devh = NULL;
 
@@ -143,37 +158,36 @@ static int find_xmos_device(unsigned int id, unsigned int pid, unsigned int list
 
 int xmos_dfu_resetdevice(void)
 {
-    libusb_control_transfer(devh, DFU_REQUEST_TO_DEV, XMOS_DFU_RESETDEVICE, 0, 0, NULL, 0, 0);
+    libusb_control_transfer(devh, USB_BMREQ_H2D_CLASS_INT, XMOS_DFU_RESETDEVICE, 0, 0, NULL, 0, 0);
     return 0;
 }
 
 int xmos_dfu_revertfactory(void)
 {
-    libusb_control_transfer(devh, DFU_REQUEST_TO_DEV, XMOS_DFU_REVERTFACTORY, 0, 0, NULL, 0, 0);
+    libusb_control_transfer(devh, USB_BMREQ_H2D_VENDOR_INT, XMOS_DFU_REVERTFACTORY, 0, 0, NULL, 0, 0);
     return 0;
 }
 
 int xmos_dfu_resetintodfu(unsigned int interface)
 {
-    libusb_control_transfer(devh, DFU_REQUEST_TO_DEV, XMOS_DFU_RESETINTODFU, 0, interface, NULL, 0, 0);
+    libusb_control_transfer(devh, USB_BMREQ_H2D_CLASS_INT, XMOS_DFU_RESETINTODFU, 0, interface, NULL, 0, 0);
     return 0;
 }
 
 int xmos_dfu_resetfromdfu(unsigned int interface)
 {
-    libusb_control_transfer(devh, DFU_REQUEST_TO_DEV, XMOS_DFU_RESETFROMDFU, 0, interface, NULL, 0, 0);
+    libusb_control_transfer(devh, USB_BMREQ_H2D_CLASS_INT, XMOS_DFU_RESETFROMDFU, 0, interface, NULL, 0, 0);
     return 0;
 }
 
 int dfu_detach(unsigned int interface, unsigned int timeout)
 {
-    libusb_control_transfer(devh, DFU_REQUEST_TO_DEV, DFU_DETACH, timeout, interface, NULL, 0, 0);
-    return 0;
+    return libusb_control_transfer(devh, USB_BMREQ_H2D_CLASS_INT, DFU_DETACH, timeout, interface, NULL, 0, dfu_timeout);
 }
 
 int dfu_getState(unsigned int interface, unsigned char *state)
 {
-    libusb_control_transfer(devh, DFU_REQUEST_FROM_DEV, DFU_GETSTATE, 0, interface, state, 1, 0);
+    libusb_control_transfer(devh, USB_BMREQ_D2H_CLASS_INT, DFU_GETSTATE, 0, interface, state, 1, 0);
     return 0;
 }
 
@@ -181,7 +195,7 @@ int dfu_getStatus(unsigned int interface, unsigned char *state, unsigned int *ti
                   unsigned char *nextState, unsigned char *strIndex)
 {
     unsigned int data[2];
-    libusb_control_transfer(devh, DFU_REQUEST_FROM_DEV, DFU_GETSTATUS, 0, interface, (unsigned char *)data, 6, 0);
+    libusb_control_transfer(devh, USB_BMREQ_D2H_CLASS_INT, DFU_GETSTATUS, 0, interface, (unsigned char *)data, 6, 0);
 
     *state = data[0] & 0xff;
     *timeout = (data[0] >> 8) & 0xffffff;
@@ -192,28 +206,13 @@ int dfu_getStatus(unsigned int interface, unsigned char *state, unsigned int *ti
 
 int dfu_clrStatus(unsigned int interface)
 {
-    libusb_control_transfer(devh, DFU_REQUEST_TO_DEV, DFU_CLRSTATUS, 0, interface, NULL, 0, 0);
+    libusb_control_transfer(devh, USB_BMREQ_H2D_CLASS_INT, DFU_CLRSTATUS, 0, interface, NULL, 0, 0);
     return 0;
 }
 
 int dfu_abort(unsigned int interface)
 {
-    libusb_control_transfer(devh, DFU_REQUEST_TO_DEV, DFU_ABORT, 0, interface, NULL, 0, 0);
-    return 0;
-}
-
-
-int xmos_dfu_save_state(unsigned int interface)
-{
-    libusb_control_transfer(devh, DFU_REQUEST_TO_DEV, XMOS_DFU_SAVESTATE, 0, interface, NULL, 0, 0);
-    printf("Save state command sent\n");
-    return 0;
-}
-
-int xmos_dfu_restore_state(unsigned int interface)
-{
-    libusb_control_transfer(devh, DFU_REQUEST_TO_DEV, XMOS_DFU_RESTORESTATE, 0, interface, NULL, 0, 0);
-    printf("Restore state command sent\n");
+    libusb_control_transfer(devh, USB_BMREQ_H2D_CLASS_INT, DFU_ABORT, 0, interface, NULL, 0, 0);
     return 0;
 }
 
@@ -221,14 +220,14 @@ unsigned int dfu_download(unsigned int interface, unsigned int block_num, unsign
 {
     //printf("... Downloading block number %d size %d\r", block_num, size);
     /* Returns actual data size transferred */
-    unsigned int transfered = libusb_control_transfer(devh, DFU_REQUEST_TO_DEV, DFU_DNLOAD, block_num, interface, data, size, 0);
+    unsigned int transfered = libusb_control_transfer(devh, USB_BMREQ_H2D_CLASS_INT, DFU_DNLOAD, block_num, interface, data, size, 0);
     return transfered;
 }
 
 int dfu_upload(unsigned int interface, unsigned int block_num, unsigned int size, unsigned char*data)
 {
     unsigned int numBytes = 0;
-    numBytes = libusb_control_transfer(devh, DFU_REQUEST_FROM_DEV, DFU_UPLOAD, block_num, interface, (unsigned char *)data, size, 0);
+    numBytes = libusb_control_transfer(devh, USB_BMREQ_D2H_CLASS_INT, DFU_UPLOAD, block_num, interface, (unsigned char *)data, size, 0);
     return numBytes;
 }
 
@@ -303,8 +302,20 @@ int write_dfu_image(char *file)
             return -1;
 
         }
-        dfu_getStatus(0, &dfuState, &timeout, &nextDfuState, &strIndex);
-        dfuBlockCount++;
+        do {
+            dfu_getStatus(0, &dfuState, &timeout, &nextDfuState, &strIndex);
+            if(nextDfuState == DFU_STATE_dfuERROR)
+            {
+                fprintf(stderr,"Error: dfu_getStatus() returned state as DFU_STATE_dfuERROR.\n");
+                return -1;
+            }
+            if(nextDfuState == DFU_STATE_dfuDNLOAD_IDLE)
+            {
+                dfuBlockCount++;
+                break;
+            }
+            Sleep(timeout);
+        }while(1);
     }
 
     if (remainder)
@@ -425,8 +436,6 @@ int main(int argc, char **argv)
     unsigned int download = 0;
     unsigned int upload = 0;
     unsigned int revert = 0;
-    unsigned int save = 0;
-    unsigned int restore = 0;
     unsigned int listdev = 0;
 
     char *firmware_filename = NULL;
@@ -480,14 +489,6 @@ int main(int argc, char **argv)
     {
         revert = 1;
     }
-    else if(strcmp(command, "--savecustomstate") == 0)
-    {
-        save = 1;
-    }
-    else if(strcmp(command, "--restorecustomstate") == 0)
-    {
-        restore = 1;
-    }
     else
     {
         print_usage(program_name,  "Invalid option passed to dfu application");
@@ -516,20 +517,15 @@ int main(int argc, char **argv)
     printf("XMOS DFU application started - Interface %d claimed\n", XMOS_DFU_IF);
 #endif
 
-    /* Dont go into DFU mode for save/restore */
-    if(save)
-    {
-        xmos_dfu_save_state(XMOS_DFU_IF);
-    }
-    else if(restore)
-    {
-        xmos_dfu_restore_state(XMOS_DFU_IF);
-    }
-    else if(!listdev)
+    if(!listdev)
     {
 #ifndef START_IN_DFU
         printf("Detaching device from application mode.\n");
-        xmos_dfu_resetintodfu(XMOS_DFU_IF);
+        if(dfu_detach(XMOS_DFU_IF, 1000) < 0)
+        {
+            fprintf(stderr, "error detaching\n");
+            return -1;
+        }
 
         libusb_release_interface(devh, XMOS_DFU_IF);
         libusb_close(devh);
@@ -580,12 +576,20 @@ int main(int argc, char **argv)
         if (download)
         {
             write_dfu_image(firmware_filename);
-            xmos_dfu_resetfromdfu(XMOS_DFU_IF);
+            if(dfu_detach(XMOS_DFU_IF, 1000) < 0)
+            {
+                fprintf(stderr, "error detaching\n");
+                return -1;
+            }
         }
         else if (upload)
         {
             read_dfu_image(firmware_filename);
-            xmos_dfu_resetfromdfu(XMOS_DFU_IF);
+            if(dfu_detach(XMOS_DFU_IF, 1000) < 0)
+            {
+                fprintf(stderr, "error detaching\n");
+                return -1;
+            }
         }
         else if (revert)
         {
@@ -593,11 +597,19 @@ int main(int argc, char **argv)
             xmos_dfu_revertfactory();
             // Give device time to revert firmware
             Sleep(2 * 1000);
-            xmos_dfu_resetfromdfu(XMOS_DFU_IF);
+            if(dfu_detach(XMOS_DFU_IF, 1000) < 0)
+            {
+                fprintf(stderr, "error detaching\n");
+                return -1;
+            }
         }
         else
         {
-            xmos_dfu_resetfromdfu(XMOS_DFU_IF);
+            if(dfu_detach(XMOS_DFU_IF, 1000) < 0)
+            {
+                fprintf(stderr, "error detaching\n");
+                return -1;
+            }
         }
 
         printf("... Returning device to application mode\n");